--- Haikara/Assets/Haikara/Runtime/Bindable/TreeViewItemsSource.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: afb59a66c60d4531b87cdd8c151eb09b
-timeCreated: 1759824170
