+++ Haikara/Assets/Haikara/Runtime/Bindable/TreeViewProperty.cs
@@ -3,28 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using Haikara.Runtime.Catalog;
 using Haikara.Runtime.View;
-using Haikara.Runtime.ViewModel;
 using Unity.Properties;
 using UnityEngine.UIElements;
 
 namespace Haikara.Runtime.Bindable
 {
-    public abstract class TreeViewEntryViewModel<T> : ViewModelBase where T : TreeViewEntryViewModel<T>
-    {
-        public List<T> Children = new();
-
-        public TreeViewItemData<T> CreateTreeViewItemData(ref int id)
-        {
-            var children = new List<TreeViewItemData<T>>(Children.Count);
-            foreach (var child in Children)
-            {
-                children.Add(child.CreateTreeViewItemData(ref id));
-            }
-            
-            return new TreeViewItemData<T>(id++, (T)this, children);   
-        }
-    }
-    
     public class TreeViewProperty<TTreeViewEntryViewModel> : CollectionViewProperty<TreeView> where TTreeViewEntryViewModel : TreeViewEntryViewModel<TTreeViewEntryViewModel>
     {
         public delegate IList<TTreeViewEntryViewModel> GetTreeViewRootItemDelegate(object? dataSource);
